requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
boto==2.48.0
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
Cython==0.27.3
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.1.1
gensim==3.1.0
html5lib==0.9999999
idna==2.6
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.5
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.0.9
Markdown==2.6.9
MarkupSafe==1.0
mistune==0.8.1
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.1
numpy==1.13.3
olefile==0.44
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
scandir==1.6
scipy==1.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.3
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
terminado==0.7
testpath==0.3.1
torch==0.2.0.post3
torchvision==0.1.9
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.8
